Update checkout recipe to include new jetId module and updated MuonSS. Adapted muonScaleResProducer.py
